Tentado com select ALL_INDEXES ao invés de Exception

+++ siga-cp/src/main/resources/db/migration/CORPORATIVO_UTF8_V114_0__Alterado_Quantidade_De_Caracteres.sql
@@ -1,34 +1,49 @@
 -- -----------------------------------------------------------------------------------------------
---	SCRIPT: ALTERA A QUANTIDADE DE CAMPOS SIGLA DE 20 PARA 30.
+--  SCRIPT: ALTERA A QUANTIDADE DE CAMPOS SIGLA DE 20 PARA 30.
 -- -----------------------------------------------------------------------------------------------
-
 DECLARE
-   index_not_exists EXCEPTION;
-   PRAGMA EXCEPTION_INIT (index_not_exists, -1418);
+    index_count INTEGER;
 BEGIN
-   EXECUTE IMMEDIATE 'DROP INDEX CORPORATIVO.DP_LOTACAO_IDX_006';
-EXCEPTION
-   WHEN index_not_exists
-   THEN
-      NULL;
+    SELECT Count(1)
+    INTO   index_count
+    FROM   all_indexes
+    WHERE  index_name = 'DP_LOTACAO_IDX_006';
+
+    IF index_count > 0 THEN
+      EXECUTE IMMEDIATE 'DROP INDEX CORPORATIVO.DP_LOTACAO_IDX_006';
+    END IF;
 END;
 /
 
 DECLARE
-   index_not_exists EXCEPTION;
-   PRAGMA EXCEPTION_INIT (index_not_exists, -1418);
+    index_count INTEGER;
 BEGIN
-   EXECUTE IMMEDIATE 'ALTER TABLE CORPORATIVO.DP_LOTACAO DROP CONSTRAINT SIGLA_LOTACAO_DP_LOTACAO_UK';
-   EXECUTE IMMEDIATE 'DROP INDEX CORPORATIVO.SIGLA_LOTACAO_DP_LOTACAO_UK';
-EXCEPTION
-   WHEN index_not_exists
-   THEN
-      NULL;
+    SELECT Count(1)
+    INTO   index_count
+    FROM   all_indexes
+    WHERE  index_name = 'SIGLA_LOTACAO_DP_LOTACAO_UK';
+
+    IF index_count > 0 THEN
+      EXECUTE IMMEDIATE
+'ALTER TABLE CORPORATIVO.DP_LOTACAO DROP CONSTRAINT SIGLA_LOTACAO_DP_LOTACAO_UK'
+  ;
+
+  EXECUTE IMMEDIATE 'DROP INDEX CORPORATIVO.SIGLA_LOTACAO_DP_LOTACAO_UK';
+END IF;
 END;
 /
 
-ALTER TABLE CORPORATIVO.DP_LOTACAO MODIFY (SIGLA_LOTACAO VARCHAR2(30));
+ALTER TABLE corporativo.dp_lotacao
+  MODIFY (sigla_lotacao VARCHAR2(30));
 
-CREATE INDEX CORPORATIVO.DP_LOTACAO_IDX_006 ON CORPORATIVO.DP_LOTACAO (ID_ORGAO_USU ASC, SIGLA_LOTACAO ASC, ID_LOTACAO ASC);
-CREATE UNIQUE INDEX CORPORATIVO.SIGLA_LOTACAO_DP_LOTACAO_UK ON CORPORATIVO.DP_LOTACAO (SIGLA_LOTACAO ASC, ID_ORGAO_USU ASC, DATA_FIM_LOT ASC);
-ALTER TABLE "CORPORATIVO"."DP_LOTACAO" ADD CONSTRAINT "SIGLA_LOTACAO_DP_LOTACAO_UK" UNIQUE ("SIGLA_LOTACAO", "ID_ORGAO_USU", "DATA_FIM_LOT") ENABLE;
+CREATE INDEX corporativo.dp_lotacao_idx_006
+  ON corporativo.dp_lotacao (id_orgao_usu ASC, sigla_lotacao ASC, id_lotacao ASC
+);
+
+CREATE UNIQUE INDEX corporativo.sigla_lotacao_dp_lotacao_uk
+  ON corporativo.dp_lotacao (sigla_lotacao ASC, id_orgao_usu ASC, data_fim_lot
+ASC);
+
+ALTER TABLE "CORPORATIVO"."dp_lotacao"
+  ADD CONSTRAINT "SIGLA_LOTACAO_DP_LOTACAO_UK" UNIQUE ("sigla_lotacao",
+  "id_orgao_usu", "data_fim_lot") ENABLE;
